+++ src/ft_convert/ft_ntoa_base.c
@@ -6,7 +6,7 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/07 18:39:26 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/08/24 23:55:54 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/10/05 00:43:57 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	ft_ntoa_base(char *b, long long n, size_t len, const char *base)
 	long long	base_len;
 	long long	i;
 
-	base_len = ft_strlen(base);
+	base_len = (long long)ft_strlen(base);
 	if (n < 0)
 		*b++ = '-';
 	while (len > 0)
