+++ src/ft_cstr/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/06 12:09:27 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/08/24 16:50:39 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/10/05 00:54:54 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ char	*ft_substr(const char *s, unsigned int start, size_t len)
 	if (strlen > start)
 		newlen = ft_strnlen(s + start, len);
 	else
-		start = strlen;
+		start = (unsigned int)strlen;
 	substr = ft_calloc(newlen + 1, sizeof(char));
 	if (!substr)
 		return (NULL);
