+++ src/ft_hash/ft_hash_combine.c
@@ -6,7 +6,7 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/28 10:39:13 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/09/28 10:41:48 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/10/05 00:55:41 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 void	ft_hash_combine(size_t *seed, char byte)
 {
-	*seed ^= byte + 0x9E3779B9 + (*seed << 6) + (*seed >> 2);
+	*seed ^= (unsigned int)byte + 0x9E3779B9 + (*seed << 6) + (*seed >> 2);
 }
